backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow-cpu
pillow==9.5.0
numpy==1.24.3
python-dotenv==1.0.0
